push code for instance create into install.node.sh

+++ install.nodes.sh
@@ -1,25 +1,44 @@
 #!/bin/bash
-KEYFILE=/Users/phagedorn/.ssh/namics.pem
-REMOTEUSER=ubuntu
+KEYFILE="/Users/phagedorn/.ssh/namics.pem"
+REMOTEUSER="ubuntu"
 instance_count=1
-rm ec2hosts.txt
+SSH_PORT=22
+secGroup="sg-1a79946d"
+ami="ami-09e3dc7d"
+keyName="namics"
+keyLocation=$KEYFILE
+region=EU-WEST-1
 
-for (( c=1; c<=instance_count; c++ ))
-do
-	echo Create Instance Number $c ...
-	bash create_instance.sh	
-done
+if [ "$1" = "debug" ]; then
+    echo DEBUG MODEUS
+    echo we will use preconfigured virtualbox instance
+    SSH_PORT="2222"
+    
+else
+    rm ec2hosts.txt
+
+    for (( c=1; c<=instance_count; c++ ))
+    do
+            echo Create Instance Number $c ...
+            instance_id=$(ec2-run-instances --region=$region -k $keyName -g $secGroup -t m1.large $ami | awk '/INSTANCE/{print $2}')
+            echo $instance_id
+            sleep 50
+            name=$(ec2-describe-instances --region=$region $instance_id | awk '/INSTANCE/{print $4}')
+            echo $name
+            echo $name >> ec2hosts.txt
+    done
+fi
 
 node=0
 
-for h in `cat ec2hosts.txt`
+for host in `cat ec2hosts.txt`
 do
-# ssh -n -i $KEYFILE $SSHOPT $REMOTEUSER@$h "sudo mysqladmin shutdown"
  let "node+=1"
  if [ $node == 1 ]; then
-   IPNODE=$(ssh -i $KEYFILE $SSHOPT $REMOTEUSER@$h "/sbin/ifconfig eth0 " | grep 'inet addr:' | cut -d: -f2 | awk '{ print $1}')
+   IPNODE=$(ssh -p $SSH_PORT -i $KEYFILE $SSHOPT $REMOTEUSER@$host "/sbin/ifconfig eth0 " | grep 'inet addr:' | cut -d: -f2 | awk '{ print $1}')
  fi
- scp -i $KEYFILE $SSHOPT install_debs.sh $REMOTEUSER@$h:/tmp
- ssh -n -i $KEYFILE $SSHOPT $REMOTEUSER@$h bash /tmp/install_debs.sh $node $IPNODE
+ scp -P $SSH_PORT $KEYFILE $SSHOPT install_debs.sh $REMOTEUSER@$host:/tmp
+ ssh -n -i $KEYFILE $SSHOPT $REMOTEUSER@$host bash /tmp/install_debs.sh $node $IPNODE
 done
+
